agents/pyproject.toml
[project]
name = "hackeurope26-agents"
version = "0.1.0"
description = "AI agent pipelines for HackEurope26"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.0",
    "google-auth>=2.35.0",
    "google-api-python-client>=2.150.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "composio>=0.7.0",
    "composio-anthropic>=0.7.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agents"]
